extract style
