+++ src/taskpane/components/ReferenceList.style.tsx
@@ -1,0 +1,11 @@
+import { mergeStyleSets } from '@fluentui/react';
+
+const classNames = mergeStyleSets({
+  container: {
+    flexDirection: 'row',
+    overflow: 'auto',
+    flex: '1 1 auto',
+  },
+});
+
+export default classNames;
