+++ src/taskpane/components/ReferenceList.tsx
@@ -1,7 +1,8 @@
 /* eslint-disable no-unused-vars */
-import { mergeStyleSets, FocusZone, FocusZoneDirection, List } from '@fluentui/react';
+import { FocusZone, FocusZoneDirection, List } from '@fluentui/react';
 import { CitationItem, MetaData } from 'citeproc';
 import React, { ReactElement } from 'react';
+import classNames from './ReferenceList.style';
 import ReferenceView from './ReferenceView';
 
 export interface bib extends MetaData, CitationItem {
@@ -11,13 +12,6 @@ export interface bib extends MetaData, CitationItem {
 interface ReferenceListProps {
   referenceList: Array<MetaData>;
 }
-const classNames = mergeStyleSets({
-  container: {
-    flexDirection: 'row',
-    overflow: 'auto',
-    flex: '1 1 auto',
-  },
-});
 
 function ReferenceList({ referenceList }: ReferenceListProps): ReactElement {
   return (
